Create Fuel_Gauge
This project reads the sensor for a 35 litre fuel tank and displays it on an OLED. A button next to the gauge loads the current fuel reading into eeprom so that it can be recalled at any later stage - even after the arduino has lost power.  This is to allow me to register the fuel onboard prior to the start of a race and then after the race is finished, be able to calculate the fuel used during the race by mentally subtracting the finish fuel level from the starting fuel level...

The display update frequency is high for 3 minutes after ignition is on.  This is so I can get a timely update of the fuel level as I'm putting fuel into the tank.  After the refuelling, the update happens infrequently as its both wildly innaccurate and distracting.It's wildly innaccurate as the fuel sloshes around crazily during a race and the highly smoothed average updated every 30secs is the only fair representation of the amount of fuel there is in the tank.
- Jos Kroon